refactor: extract __debugInfo() credential redaction into a trait
The same 12-line __debugInfo() override that masks credential
attributes lives on two unrelated base classes: Item\Configuration\Base
and Item\AuthenticationMethod\Base. Extract the shared logic and
the $sensitiveAttributes property into a Didww\Traits\Redacts-
SensitiveAttributes trait, mirroring the existing Didww\Traits\
HasSafeAttributes / HasEnumAttributes pattern. Both base classes now
`use RedactsSensitiveAttributes` and the duplicate is gone.

Behaviour is unchanged — the trait's `__debugInfo()` body is
character-for-character identical to the previous inline methods, and
the existing redaction tests on both base classes continue to pass.
