+++ src/Item/Configuration/Base.php
@@ -12,24 +12,15 @@ use Didww\Enum\TransportProtocol;
 use Didww\Enum\TxDtmfFormat;
 use Didww\Traits\HasEnumAttributes;
 use Didww\Traits\HasSafeAttributes;
+use Didww\Traits\RedactsSensitiveAttributes;
 
 abstract class Base
 {
     use HasSafeAttributes;
     use HasEnumAttributes;
+    use RedactsSensitiveAttributes;
 
     protected $attributes = [];
-
-    /**
-     * Attribute names whose values are credentials. The wire format is
-     * unchanged — toJsonApiArray() still emits the real values — but
-     * __debugInfo() / __toString() redact these so var_dump / print_r /
-     * default error reports never leak the credential downstream.
-     * Subclasses extend this array.
-     *
-     * @var string[]
-     */
-    protected array $sensitiveAttributes = [];
 
     public function __construct($attributes = [])
     {
@@ -42,22 +33,6 @@ abstract class Base
             'type' => $this->getType(),
             'attributes' => $this->getAttributes(),
         ];
-    }
-
-    /**
-     * Custom var_dump / print_r output: redact sensitive attribute values
-     * so credentials never leak through default debugging surfaces.
-     */
-    public function __debugInfo(): array
-    {
-        $masked = [];
-        foreach ($this->attributes as $key => $value) {
-            $masked[$key] = (in_array($key, $this->sensitiveAttributes, true) && null !== $value)
-                ? '[FILTERED]'
-                : $value;
-        }
-
-        return ['type' => $this->getType(), 'attributes' => $masked];
     }
 
     public function getAttributes()
